atividade 01

+++ AtvI/src/atvi/Problema.java
@@ -4,7 +4,7 @@ public class Problema {
 
     //Funcao de avaliacao - funcao de rastringin
     public void calcularFuncaoObjetivo(Individuo individuo) {
-        Double soma = (double) 10 * individuo.getnVar();
+        Double soma = 10.0 * individuo.getnVar();
 
         for (Double var : individuo.getVariaveis()) {
             soma += Math.pow(var, 2) - (10 * Math.cos(2 * Math.PI * var));
